perfected db,
everything works fine till here

+++ app.py
@@ -3,7 +3,7 @@ from functools import wraps
 from flask import Flask, render_template, redirect, url_for, flash, abort, send_from_directory
 from flask_bootstrap import Bootstrap
 from flask_ckeditor import CKEditor
-from datetime import date
+from datetime import date, datetime
 from flask_login import LoginManager, current_user, login_required
 from flask_gravatar import Gravatar
 import secrets
@@ -31,6 +31,7 @@ Bootstrap(app)
 
 # CONNECT TO DB
 app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///blog.db'
+# app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///blogs.db'
 app.config['SQLALCHEMY_TRACK_MODIFICATIONS'] = False
 UPLOAD_USER_IMG = "dynamic/profile_pic"
 app.config['UPLOAD_USER_IMG'] = UPLOAD_USER_IMG
@@ -126,12 +127,13 @@ def show_post(post_id):
             )
             db.session.add(new_comment)
             db.session.commit()
-            return redirect(url_for('get_all_posts'))
-
-        img_url = url_for('uploaded_file', filename=requested_post.img_url)
+            return redirect(url_for('show_post', post_id=requested_post.id))
+        img_url = None
+        if requested_post.img_url:
+            img_url = url_for('uploaded_file', filename=requested_post.img_url)
 
         return render_template("post.html", post=requested_post, form=form, current_user=current_user, img_url=img_url,
-                               year=year)
+                               year=year, auth_users=auth_users)
     flash("You're not logged-in. Kindly Log-in")
     return redirect(url_for('get_all_posts'))
 
@@ -140,28 +142,79 @@ def show_post(post_id):
 @login_required
 def add_new_post():
     form = CreatePostForm()
+    img_url = None
     if form.validate_on_submit():
-        img = form.img_url
-        img_id = img_to_uuid(img)
-        img.data.save(os.path.join(app.config['UPLOAD_BLOG_IMG'], img_id))
-        new_post = BlogPost(
-            title=form.title.data,
-            subtitle=form.subtitle.data,
-            body=form.body.data,
-            img_url=img_id,
-            author=current_user,
-            date=date.today().strftime("%B %d, %Y")
-        )
-        db.session.add(new_post)
-        db.session.commit()
-        return redirect(url_for("get_all_posts"))
-    return render_template("make-post.html", form=form, year=year)
+        if form.img_url.data:
+            img = form.img_url
+            img_id = img_to_uuid(img)
+            img.data.save(os.path.join(app.config['UPLOAD_BLOG_IMG'], img_id))
+            img_url = img_id
+
+        try:
+            print("try")
+            new_post = BlogPost(
+                title=form.title.data,
+                subtitle=form.subtitle.data,
+                body=form.body.data,
+                img_url=img_url,
+                author=current_user,
+                date=datetime.now().strftime("%B %d, %Y %I:%M %p")
+            )
+            print("new post")
+            db.session.add(new_post)
+            print("added")
+            db.session.commit()
+            print("commit")
+            return redirect(url_for("get_all_posts"))
+        except Exception as e:
+            db.session.rollback()
+            app.logger.error(str(e))
+            form.errors['submit'] = 'Failed to create new post.'
+    return render_template("make-post.html", form=form)
+#
+# @app.route("/new-post", methods=["GET", "POST"])
+# @login_required
+# def add_new_post():
+#     form = CreatePostForm()
+#     img_url = None
+#     if form.validate_on_submit():
+#         if form.img_url.data:
+#             img = form.img_url
+#             img_id = img_to_uuid(img)
+#             img.data.save(os.path.join(app.config['UPLOAD_BLOG_IMG'], img_id))
+#             img_url = img_id
+#         else:
+#             img_url = None
+#
+#     if form.validate_on_submit():
+#         try:
+#             new_post = BlogPost(
+#                 title=form.title.data,
+#                 subtitle=form.subtitle.data,
+#                 body=form.body.data,
+#                 img_url=img_url,
+#                 author=current_user,
+#                 date=datetime.now().strftime("%B %d, %Y %I:%M %p")
+#             )
+#             db.session.add(new_post)
+#             db.session.commit()
+#             return redirect(url_for("get_all_posts"))
+#         except Exception as e:
+#             db.session.rollback()
+#             app.logger.error(str(e))
+#             form.errors['submit'] = 'Failed to create new post.'
+#
+#     return render_template("make-post.html", form=form)
 
 
 @app.route("/edit-post/<int:post_id>", methods=["GET", "POST"])
 @login_required
 def edit_post(post_id):
     post = BlogPost.query.get_or_404(post_id)
+    file_path = None
+    img_id = None
+    if post.img_url:
+        file_path = os.path.join(UPLOAD_BLOG_IMG, post.img_url)
     edit_form = CreatePostForm(
         title=post.title,
         subtitle=post.subtitle,
@@ -169,20 +222,24 @@ def edit_post(post_id):
         author=post.author,
         body=post.body
     )
-    file_path = os.path.join(UPLOAD_BLOG_IMG, post.img_url)
     if edit_form.validate_on_submit():
-        img = edit_form.img_url
-        img_id = img_to_uuid(img)
-        img.data.save(os.path.join(app.config['UPLOAD_BLOG_IMG'], img_id))
-
-        post.title = edit_form.title.data
-        post.subtitle = edit_form.subtitle.data
-        post.img_url = img_id
-        post.body = edit_form.body.data
-        db.session.commit()
-
-        delete_file(file_path)
-        return redirect(url_for("show_post", post_id=post_id))
+        if edit_form.img_url:
+            img = edit_form.img_url
+            img_id = img_to_uuid(img)
+            img.data.save(os.path.join(app.config['UPLOAD_BLOG_IMG'], img_id))
+        try:
+            post.title = edit_form.title.data
+            post.subtitle = edit_form.subtitle.data
+            post.img_url = img_id
+            post.body = edit_form.body.data
+            db.session.commit()
+            if file_path and img_id:
+                delete_file(file_path)
+            return redirect(url_for("show_post", post_id=post_id))
+        except Exception as e:
+            db.session.rollback()
+            app.logger.error(str(e))
+            edit_form.errors['submit'] = 'Failed to create new post.'
 
     return render_template("make-post.html", form=edit_form, year=year)
 
@@ -191,17 +248,35 @@ def edit_post(post_id):
 @login_required
 def delete_post(post_id):
     post_to_delete = BlogPost.query.get_or_404(post_id)
-    file_path = os.path.join(UPLOAD_BLOG_IMG, post_to_delete.img_url)
+    file_path = None
+    if post_to_delete.img_url:
+        file_path = os.path.join(UPLOAD_BLOG_IMG, post_to_delete.img_url)
     try:
         Comment.query.filter_by(post_id=post_id).delete()
         db.session.delete(post_to_delete)
         db.session.commit()
-        delete_file(file_path)
+        if post_to_delete.img_url:
+            delete_file(file_path)
         flash("Blog post was deleted.")
         return redirect(url_for('get_all_posts'))
     except:
         flash("Whoops!! There was a problem deleting that post.")
         return redirect(url_for('get_all_posts'))
+
+
+@app.route("/delete_comment/<int:cmt_id>", methods=["GET", "POST"])
+@login_required
+def delete_cmt(cmt_id):
+    cmt_to_delete = Comment.query.get_or_404(cmt_id)
+    try:
+        Comment.query.filter_by(post_id=cmt_id).delete()
+        db.session.delete(cmt_to_delete)
+        db.session.commit()
+        flash("Comment was deleted.")
+        return redirect(url_for('show_post', post_id=cmt_to_delete.post_id))
+    except:
+        flash("Whoops!! There was a problem deleting that comment.")
+        return redirect(url_for('show_post', post_id=cmt_to_delete.post_id))
 
 
 @app.errorhandler(404)
@@ -214,6 +289,16 @@ def page_not_found(e):
     return render_template("500.html"), 500
 
 
+@app.errorhandler(401)
+def page_not_found(e):
+    return render_template("401.html"), 401
+
+
+@app.errorhandler(403)
+def page_not_found(e):
+    return render_template("403.html"), 403
+
+
 if __name__ == "__main__":
-    app.run(host='0.0.0.0', debug=True)
-# serve(app, host='0.0.0.0', port=80, threads=1)
+    app.run(debug=True)
+
